+++ Changelog.md
@@ -5,6 +5,7 @@
 * added perfect power factor to math
 * generalized quantum phase estimation and integrated in quantum-period-finding
 * fixed coprime selection in shor
+* added prime check in shor
 
 ## Version 0.1.0
 * implemented quantom states, quantum gates, gate registry and quantum cirquits